Fix gosec G115 integer overflow findings in unix auth transport
golangci-lint flagged three unchecked integer conversions introduced by
the peer credential check.

Take the peer uid as an int64 in verifySocketPeerUID rather than a
uint32. The platform credential structs use uint32, which widens to
int64 losslessly, and os.Geteuid returns an int that also widens
losslessly, so the comparison no longer needs a signed/unsigned
conversion in either the callers or the test.

The remaining conversion, uintptr to int for the descriptor handed to
GetsockoptUcred/GetsockoptXucred, cannot be avoided because the unix
wrappers take an int. A descriptor supplied by RawConn.Control is always
a small non-negative value, so it is annotated with a documented nolint
matching the convention already used for the Win32 ACE cast.

The darwin file carries the same two conversions as the linux file and is
fixed alongside it, even though CI only lints linux and windows.

Verified with golangci-lint v2.11.4 (the version CI pins) against GOOS
linux, windows, and darwin, and by running the auth transport tests on
linux, where the full round trip exercises SO_PEERCRED end to end.

Copilot-Session: 4a9b964f-5a93-4c5e-b779-4ad3572fc658

+++ cli/azd/cmd/auth_transport_peer_darwin.go
@@ -27,6 +27,9 @@ func verifySocketPeer(conn *net.UnixConn) error {
 		credErr error
 	)
 	if err := rawConn.Control(func(fd uintptr) {
+		// A descriptor supplied by RawConn.Control is always a small
+		// non-negative value, so the conversion cannot overflow.
+		//nolint:gosec // G115: file descriptors always fit in an int.
 		cred, credErr = unix.GetsockoptXucred(int(fd), unix.SOL_LOCAL, unix.LOCAL_PEERCRED)
 	}); err != nil {
 		return fmt.Errorf("accessing AZD_AUTH_ENDPOINT socket handle: %w", err)
@@ -34,5 +37,5 @@ func verifySocketPeer(conn *net.UnixConn) error {
 	if credErr != nil {
 		return fmt.Errorf("reading AZD_AUTH_ENDPOINT socket peer credentials: %w", credErr)
 	}
-	return verifySocketPeerUID(cred.Uid)
+	return verifySocketPeerUID(int64(cred.Uid))
 }
